Getting Somewhere...
I'm finally starting to see where this particular fix will end up.  It needs to be sorted on priority, though, so I think a nice way to do this would be to turn the Accept header into a binary tree, with each content type taking up a node, sorted from highest priority (on the left) to lowest priority (on the right).